requirements.txt
# Requirements for Multi-Target Face Detection and Recognition System

# Core dependencies
opencv-python>=4.8.0
insightface>=0.7.3
numpy>=1.24.0
onnxruntime>=1.15.0

# For Azure blob storage integration
requests>=2.31.0
azure-storage-blob>=12.19.0

# Environment variables management
python-dotenv>=1.0.0

# Optional dependencies for enhanced functionality
Pillow>=10.0.0
matplotlib>=3.7.0

# For video processing
imageio>=2.31.0

# For data handling
pickle-mixin>=1.0.2

# Development and testing (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Note: The system will automatically download the InsightFace models on first run
# Models are stored in ~/.insightface/models/ directory